feat(storage)!: migrate lease/state store from SQLite to PostgreSQL
dora is now Postgres-only. Lease and management state (leases v4/v6, runtime
reservations, operations/audit, config candidates) lives in PostgreSQL instead
of an embedded SQLite file.

Storage layer:
- rewrite the migrations as Postgres DDL (BIGINT for v4 ip/network + epoch
  timestamps, BYTEA for v6 addrs/ids, native BOOLEAN, ON CONFLICT upserts).
- port ip-manager: SqliteDb -> PostgresDb (PgPool), all 44 queries to the
  Postgres dialect ($N placeholders, COUNT(..) as x! non-null overrides),
  regenerate the offline sqlx-data.json (builds stay offline via SQLX_OFFLINE).
- PostgresDb::new runs the embedded migrations on startup, as before.

Config / wiring:
- DATABASE_URL is now a Postgres connection string; default
  postgres://dora:dora@localhost/dora (override in prod). main.rs uses PostgresDb.

Tests:
- in-memory SQLite is gone, so each test provisions its own transient database:
  PostgresDb::new_test() (unit tests) and create/drop_test_database() for the
  out-of-process bin integration tests. The v4 netns tests reach Postgres over
  the veth host IP (DORA_TEST_DB_HOST, default 192.168.2.99); the v6 loopback
  tests use the admin host directly.
- CI: add a postgres:16 service + DATABASE_URL to the test and coverage jobs
  (build/lint jobs stay offline and need no database).

Docs: README build/run, bin/README, docker.md, pi_setup.md updated for Postgres.

BREAKING CHANGE: dora requires a reachable PostgreSQL database; the SQLite file
backend and file-path DATABASE_URL are removed.

+++ migrations/20260708120000_operations.sql
@@ -21,9 +21,9 @@ CREATE TABLE IF NOT EXISTS operations(
     result        TEXT,
     error_code    TEXT,
     error_message TEXT,
-    created_at    INTEGER NOT NULL,
-    started_at    INTEGER,
-    completed_at  INTEGER
+    created_at    BIGINT  NOT NULL,
+    started_at    BIGINT,
+    completed_at  BIGINT
 );
 -- listing/pruning is newest-first by creation time
 CREATE INDEX IF NOT EXISTS idx_operations_created_at ON operations (created_at);
